move playground to route with id url-param
